+ Infile self-contained call implementation.
